% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{function that will output a vector of the first n terms in the child’s game Fizz Buzz}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{A real number above 0}
}
\value{
vector containing real numbner, Fizz, Buzz, FizzBuzz
}
\description{
function that will output a vector of the first n terms in the child’s game Fizz Buzz
}
\examples{
FizzBuzz(10)
}
