% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{weather information for Flagstaff from 2015 to 2019 based on records from the pulliam airport in flagstaff}
\format{
A data frame with 1719 observations of 5 Variables
\describe{
  \item{DATE}{Date of the record (YMD Format)}
  \item{PRCP}{Amount of Precipication in inches}
  \item{SNOW}{Amount of Snowfall in inches}
  \item{TMAX}{Maximum tempature in Fahrenheit}
  \item{TMAX}{Minimum tempature in Fahrenheit}
}
}
\source{
\url{https://www.ncdc.noaa.gov/}
}
\usage{
Flagstaff_Weather
}
\description{
weather information for Flagstaff from 2015 to 2019 based on records from the pulliam airport in flagstaff
}
\keyword{datasets}
